[Wsign-compare] Close Wno-error of sign-compare (#47252)
* Close Wno-error of sign-compare

* fix sign-compare in graph_node_test.cc

* fix sign-compare in graph_node_test.cc file

* fix sign-compare in common_graph_table.cc file

+++ paddle/fluid/distributed/ps/table/common_graph_table.cc
@@ -556,7 +556,7 @@ int32_t GraphTable::dump_edges_to_ssd(int idx) {
 }
 int32_t GraphTable::make_complementary_graph(int idx, int64_t byte_size) {
   VLOG(0) << "make_complementary_graph";
-  const int64_t fixed_size = byte_size / 8;
+  const size_t fixed_size = byte_size / 8;
   // std::vector<int64_t> edge_array[task_pool_size_];
   std::vector<std::unordered_map<uint64_t, int>> count(task_pool_size_);
   std::vector<std::future<int>> tasks;
